refactor: colocate HTTP client metrics helpers
