--- pkg/util/http_roundtripper_integration_test.go
@@ -1,31 +1,0 @@
-/*
-Copyright 2026 The KEDA Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package util
-
-import (
-	"fmt"
-	"testing"
-
-	. "github.com/onsi/gomega"
-)
-
-func TestCreateHTTPClient_TransportIsInstrumented(t *testing.T) {
-	RegisterTestingT(t)
-
-	client := CreateHTTPClient(0, false)
-	Expect(fmt.Sprintf("%T", client.Transport)).To(Equal("*metricscollector.InstrumentedRoundTripper"))
-}
